+++ src/index.js
@@ -1,5 +1,5 @@
-const angular = require('angular')
-const CacheFactory = require('cachefactory')
+import angular from 'angular'
+import CacheFactory from '../node_modules/cachefactory/dist/cachefactory.es2015'
 
 CacheFactory.utils.equals = angular.equals
 CacheFactory.utils.isObject = angular.isObject
@@ -23,10 +23,4 @@ angular.module('angular-cache', [])
   .provider('BinaryHeap', BinaryHeapProvider)
   .provider('CacheFactory', CacheFactoryProvider)
 
-module.exports = 'angular-cache'
-try {
-  module.exports.name = 'angular-cache'
-} catch (err) {
-
-}
-
+export default 'angular-cache'